+++ flutter_readium_platform_interface/lib/src/shared/publication/metadata/contributor.dart
@@ -86,27 +86,29 @@ class Contributor extends BaseCollection {
   }
 
   Contributor copyWith({
-    double? position,
-    LocalizedString? localizedName,
-    String? identifier,
-    List<AltIdentifier>? altIdentifiers,
-    LocalizedString? localizedSortAs,
-    List<Link>? links,
-    List<String>? roles,
-    Map<String, dynamic>? additionalProperties,
+    Object? position = unset,
+    Object? localizedName = unset,
+    Object? identifier = unset,
+    Object? altIdentifiers = unset,
+    Object? localizedSortAs = unset,
+    Object? links = unset,
+    Object? roles = unset,
+    Object? additionalProperties = unset,
   }) {
-    final mergeProperties = Map<String, dynamic>.of(this.additionalProperties)
-      ..addAll(additionalProperties ?? {})
-      ..removeWhere((key, value) => value == null);
+    final mergeProperties = copyAdditionalProperties(additionalProperties: additionalProperties);
 
     return Contributor(
-      position: position ?? this.position,
-      localizedName: localizedName ?? this.localizedName,
-      identifier: identifier ?? this.identifier,
-      altIdentifiers: altIdentifiers ?? this.altIdentifiers,
-      localizedSortAs: localizedSortAs ?? this.localizedSortAs,
-      links: links ?? this.links,
-      roles: roles ?? this.roles,
+      position: identical(position, unset) ? this.position : (position as double?)!,
+      localizedName: identical(localizedName, unset) ? this.localizedName : (localizedName as LocalizedString?)!,
+      identifier: identical(identifier, unset) ? this.identifier : (identifier as String?)!,
+      altIdentifiers: identical(altIdentifiers, unset)
+          ? this.altIdentifiers
+          : (altIdentifiers as List<AltIdentifier>?)!,
+      localizedSortAs: identical(localizedSortAs, unset)
+          ? this.localizedSortAs
+          : (localizedSortAs as LocalizedString?)!,
+      links: identical(links, unset) ? this.links : (links as List<Link>?)!,
+      roles: identical(roles, unset) ? this.roles : (roles as List<String>?)!,
       additionalProperties: mergeProperties,
     );
   }