+++ flutter_readium_platform_interface/lib/src/shared/publication/metadata/periodical.dart
@@ -91,29 +91,31 @@ class Periodical extends BaseCollection {
   }
 
   Periodical copyWith({
-    double? position,
-    LocalizedString? localizedName,
-    String? identifier,
-    List<AltIdentifier>? altIdentifiers,
-    LocalizedString? localizedSortAs,
-    List<Link>? links,
-    List<Issue>? issues,
-    List<Volume>? volumes,
-    Map<String, dynamic>? additionalProperties,
+    Object? position = unset,
+    Object? localizedName = unset,
+    Object? identifier = unset,
+    Object? altIdentifiers = unset,
+    Object? localizedSortAs = unset,
+    Object? links = unset,
+    Object? issues = unset,
+    Object? volumes = unset,
+    Object? additionalProperties = unset,
   }) {
-    final mergeProperties = Map<String, dynamic>.of(this.additionalProperties)
-      ..addAll(additionalProperties ?? {})
-      ..removeWhere((key, value) => value == null);
+    final mergeProperties = copyAdditionalProperties(additionalProperties: additionalProperties);
 
     return Periodical(
-      position: position ?? this.position,
-      localizedName: localizedName ?? this.localizedName,
-      identifier: identifier ?? this.identifier,
-      altIdentifiers: altIdentifiers ?? this.altIdentifiers,
-      localizedSortAs: localizedSortAs ?? this.localizedSortAs,
-      links: links ?? this.links,
-      issues: issues ?? this.issues,
-      volumes: volumes ?? this.volumes,
+      position: identical(position, unset) ? this.position : (position as double?)!,
+      localizedName: identical(localizedName, unset) ? this.localizedName : (localizedName as LocalizedString?)!,
+      identifier: identical(identifier, unset) ? this.identifier : (identifier as String?)!,
+      altIdentifiers: identical(altIdentifiers, unset)
+          ? this.altIdentifiers
+          : (altIdentifiers as List<AltIdentifier>?)!,
+      localizedSortAs: identical(localizedSortAs, unset)
+          ? this.localizedSortAs
+          : (localizedSortAs as LocalizedString?)!,
+      links: identical(links, unset) ? this.links : (links as List<Link>?)!,
+      issues: identical(issues, unset) ? this.issues : (issues as List<Issue>?)!,
+      volumes: identical(volumes, unset) ? this.volumes : (volumes as List<Volume>?)!,
       additionalProperties: mergeProperties,
     );
   }